splitscreen now working!
Turns out it's a lot easier to code when it's not 1am! Who woulda
thought? Just had to change triangle coords for each fractal. Much
simpler than the solution I attempted yesterday. Only problem is that
the center of the screen is still the center of the fractal. Gotta
translate it!